Add Computer.convert_pyam convenience method from message_ix.Reporter
